+++ app/src/main/java/com/gianlu/pretendyourexyzzy/api/models/cards/BaseCard.java
@@ -1,18 +1,33 @@
 package com.gianlu.pretendyourexyzzy.api.models.cards;
 
+import android.util.Log;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import java.io.Serializable;
+import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public abstract class BaseCard implements Serializable {
     private static final Pattern HTML_IMAGE_PATTERN = Pattern.compile("^<img.+src='(.*?)'.+/>$");
+    private static final String TAG = BaseCard.class.getSimpleName();
     private transient String imageUrl = null;
 
     @NonNull
     public abstract String text();
+
+    @NonNull
+    public final String textUnescaped() {
+        try {
+            return URLDecoder.decode(text(), "UTF-8");
+        } catch (UnsupportedEncodingException | IllegalArgumentException ex) {
+            Log.e(TAG, "Failed unescaping text: " + text(), ex);
+            return text();
+        }
+    }
 
     @Nullable
     public abstract String watermark();