+++ app/src/main/java/com/gianlu/pretendyourexyzzy/customdecks/view/GeneralInfoFragment.java
@@ -13,19 +13,40 @@ import androidx.annotation.Nullable;
 
 import com.gianlu.commonutils.dialogs.FragmentWithDialog;
 import com.gianlu.pretendyourexyzzy.R;
+import com.gianlu.pretendyourexyzzy.api.crcast.CrCastApi;
+import com.gianlu.pretendyourexyzzy.api.crcast.CrCastDeck;
 
 import org.jetbrains.annotations.NotNull;
+
+import xyz.gianlu.pyxoverloaded.model.UserProfile;
 
 public final class GeneralInfoFragment extends FragmentWithDialog {
 
     @NotNull
-    public static GeneralInfoFragment get(@NotNull Context context, @NotNull String name, @NotNull String watermark, @NotNull String description) {
+    public static GeneralInfoFragment get(@NotNull Context context, @NotNull UserProfile.CustomDeckWithCards deck) {
         GeneralInfoFragment fragment = new GeneralInfoFragment();
         Bundle args = new Bundle();
         args.putString("title", context.getString(R.string.info));
-        args.putString("name", name);
-        args.putString("watermark", watermark);
-        args.putString("desc", description);
+        args.putBoolean("crCast", false);
+        args.putString("name", deck.name);
+        args.putString("watermark", deck.watermark);
+        args.putString("desc", deck.desc);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @NotNull
+    public static GeneralInfoFragment get(@NotNull Context context, @NotNull CrCastDeck deck) {
+        GeneralInfoFragment fragment = new GeneralInfoFragment();
+        Bundle args = new Bundle();
+        args.putString("title", context.getString(R.string.info));
+        args.putBoolean("crCast", true);
+        args.putString("name", deck.name);
+        args.putString("watermark", deck.watermark);
+        args.putString("desc", deck.desc);
+        args.putString("lang", deck.lang);
+        args.putBoolean("private", deck.privateDeck);
+        args.putSerializable("state", deck.state);
         fragment.setArguments(args);
         return fragment;
     }
@@ -34,14 +55,38 @@ public final class GeneralInfoFragment extends FragmentWithDialog {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         LinearLayout layout = (LinearLayout) inflater.inflate(R.layout.fragment_view_custom_deck_info, container, false);
-        TextView name = layout.findViewById(R.id.viewCustomDeck_name);
-        name.setText(requireArguments().getString("name"));
+        Bundle args = requireArguments();
 
-        TextView desc = layout.findViewById(R.id.viewCustomDeck_desc);
-        desc.setText(requireArguments().getString("desc"));
+        TextView name = layout.findViewById(R.id.viewCustomDeck_name);
+        name.setText(args.getString("name"));
 
         TextView watermark = layout.findViewById(R.id.viewCustomDeck_watermark);
-        watermark.setText(requireArguments().getString("watermark"));
+        watermark.setText(args.getString("watermark"));
+
+        String descStr = args.getString("desc");
+        if (descStr == null || descStr.isEmpty()) {
+            layout.findViewById(R.id.viewCustomDeck_desc).setVisibility(View.GONE);
+            layout.findViewById(R.id.viewCustomDeck_descLabel).setVisibility(View.GONE);
+        } else {
+            TextView desc = layout.findViewById(R.id.viewCustomDeck_desc);
+            desc.setText(descStr);
+        }
+
+        LinearLayout crCast = layout.findViewById(R.id.viewCustomDeck_crCast);
+        if (args.getBoolean("crCast", false)) {
+            crCast.setVisibility(View.VISIBLE);
+
+            TextView state = crCast.findViewById(R.id.viewCustomDeck_state);
+            state.setText(((CrCastApi.State) args.getSerializable("state")).toFormal());
+
+            TextView privateDeck = crCast.findViewById(R.id.viewCustomDeck_private);
+            privateDeck.setText(args.getBoolean("private") ? R.string.yes : R.string.no);
+
+            TextView lang = crCast.findViewById(R.id.viewCustomDeck_lang);
+            lang.setText(args.getString("lang"));
+        } else {
+            crCast.setVisibility(View.GONE);
+        }
 
         return layout;
     }