+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.0] - 16-08-2020
+### Added
+- Added support for CrCast (view and use decks)
+
+### Fixed
+- Fixed crash when using some cards
+
+
+## [4.0.2] - 11-08-2020
+### Added
+- Export/share custom decks
+
+### Changed
+- Refresh game after 3 seconds it has been stuck
+
+### Fixed
+- Fixed importing cards with 0 pick
+- Fixed unescaped HTML in cards
+
+
+## [4.0.1] - 01-08-2020
+### Fixed
+- Fixed crash when joining game
+
+
 ## [4.0.0] - 28-07-2020
 ### Added
 - Highlight Overloaded users in player list